Search Form additional updates
convert ‘state field’ into a read only text input.

the reason for this being that while it should be implied that this is going to search massachusetts legislators, this particular element would be unnoticed to screen readers that are in forms mode, as all non-form elements are skipped over.  To make sure this isn’t missed, and to allow for this to ever be extended up on in the future, the state ‘label’ was converted to a read label element, and a read only text input was added, with a value set to massachusetts.  minor style revisions were made to readonly inputs for this form specifically, to better adhere to the original design.